Project_Changed

+++ app/Http/Controllers/Admin/CommentController.php
@@ -3,52 +3,40 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\CommentsRequest;
 use App\Models\Comment;
-use App\Models\User;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\View\View;
+use Illuminate\Http\Request;
 
 class CommentController extends Controller
 {
     /**
-     * Show the application comments index.
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
      */
-    public function index(): View
+    public function index()
     {
-        return view('admin.comments.index', [
-            'comments' => Comment::with('post', 'author')->latest()->paginate(50)
-        ]);
-    }
+        $comments = Comment::with('post')->paginate(10);
 
-    /**
-     * Display the specified resource edit form.
-     */
-    public function edit(Comment $comment): View
-    {
-        return view('admin.comments.edit', [
-            'comment' => $comment,
-            'users' => User::pluck('name', 'id')
-        ]);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(CommentsRequest $request, Comment $comment): RedirectResponse
-    {
-        $comment->update($request->validated());
-
-        return redirect()->route('admin.comments.edit', $comment)->withSuccess(__('comments.updated'));
+        return view('admin.comments.index', compact('comments'));
     }
 
     /**
      * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
      */
-    public function destroy(Comment $comment): RedirectResponse
+    public function destroy(Comment $comment)
     {
-        $comment->delete();
+        if ($comment->user_id != auth()->user()->id && auth()->user()->is_admin == false) {
+            flash()->overlay("You can't delete other peoples comment.");
 
-        return redirect()->route('admin.comments.index')->withSuccess(__('comments.deleted'));
+            return redirect('/admin/posts');
+        }
+
+        $comment->delete();
+        flash()->overlay('Comment deleted successfully.');
+
+        return redirect('/admin/comments');
     }
 }
